+++ Tools/ResourceEditor/Classes/SceneEditor/ComboBox.h
@@ -65,7 +65,7 @@ protected:
     
     int32 selectionIndex;
     
-    int32 listWidth;
+    float32 listWidth;
     
 };
