+++ Tools/ResourceEditor/Classes/SceneEditor/EditorSettings.cpp
@@ -204,7 +204,7 @@ void EditorSettings::AddLastOpenedFile(const String & pathToFile)
     
     filesList.insert(filesList.begin(), pathToFile);
     count = 0;
-    for(;(count < filesList.size()) && (count < RESENT_FILES_COUNT); ++count)
+    for(;(count < (int32)filesList.size()) && (count < RESENT_FILES_COUNT); ++count)
     {
         settings->SetString(Format("LastOpenedFile_%d", count), filesList[count]);
     }
